Tweaks to hopefully improve performance.
Labels now use 16 bytes instead of 8, allowing for more of them in one tree.
Unary and binary nodes now have the same probability to appear as everything else, resulting in less of them over-all.
Fixed a bug involving wrong ordering of results (based on dice index first, score second).

+++ src/trees/translate.rs
@@ -3,11 +3,11 @@ use crate::vm::definitions::{Instruction, Source, JumpCondition, RESULT_REGISTER
 use crate::vm::definitions::BinaryOperation::Set;
 
 pub fn translate_tree(tree: &Node) -> Vec<Instruction> {
-    let mut next_label = 0u8;
+    let mut next_label = 0u16;
     translate_subtree(tree, STACK_START, &mut next_label)
 }
 
-fn translate_subtree(tree: &Node, stack_ptr: u8, next_label: &mut u8) -> Vec<Instruction> {
+fn translate_subtree(tree: &Node, stack_ptr: u8, next_label: &mut u16) -> Vec<Instruction> {
     match tree {
         Node::Null => {
             // do nothing
